dont reinstall cue-gen

+++ m/defn/dev/make_crd_cue_gen.sh
@@ -3,11 +3,14 @@
 set -eufo pipefail
 
 function main {
-	go mod init meh
-	go get istio.io/tools/cmd/cue-gen
-	go install istio.io/tools/cmd/cue-gen
-	rm go.mod go.sum
-	"$(go env GOPATH)/bin/cue-gen" -f=cue.yaml --crd=true
+	if ! test -x "$(go env GOPATH || true)/bin/cue-gen"; then
+		go mod init meh
+		go get istio.io/tools/cmd/cue-gen
+		go install istio.io/tools/cmd/cue-gen
+		rm go.mod go.sum
+	fi
+
+	"$(go env GOPATH || true)/bin/cue-gen" -f=cue.yaml --crd=true
 }
 
 main "$@"
